Adjust Tipis biar lebih rapi
