+++ resources/js/Pages/Home.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, Heading, Button } from "@chakra-ui/react";
-import Footer from "@/Components/Footer";
-import Navbar from "@/Components/Navbar";
+import Footer from "@/Components/Footer/Footer";
+import Navbar from "@/Components/Navbar/Navbar";
 
 const Home = () => {
     return (
