+++ routes/web.php
@@ -24,6 +24,7 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
+// ini saya pakai untuk testing page -- kevin
 Route::get('/home', function () {
     return Inertia::render('Home');
 });
